Refactor DialogFooter and add responsive Design
